Renamed "src" to "pyndemic" and updated imports.
Moved doc assets to "doc" directory.
Added IDE configuration files to .gitignore.
