+++ pyndemic/config.py
@@ -4,7 +4,8 @@ from configparser import ConfigParser
 
 
 ROOT_DIR = os.path.join(os.path.dirname(__file__), os.pardir)
-SETTINGS_DEFAULT_LOCATION = os.path.join(ROOT_DIR, 'settings.cfg')
+WORK_DIR = os.path.dirname(__file__)
+SETTINGS_DEFAULT_LOCATION = os.path.join(WORK_DIR, 'settings.cfg')
 
 _CACHED_SETTINGS = None
