+++ pyndemic/disease.py
