+++ pyndemic/exceptions.py
