+++ pyndemic/formatter.py
