+++ pyndemic/settings.cfg
