+++ test/test_disease.py
@@ -2,7 +2,7 @@
 
 from unittest import TestCase, skip, expectedFailure
 
-from src.disease import Disease, NoHealthException
+from pyndemic.disease import Disease, NoHealthException
 
 
 class DiseaseTestCase(TestCase):
